Adding app skeleton
This is largely to setup a new Poetry environment with Black for code formatting.

The README is also updated with a basic structure and context.

Note: This skeleton will improve in future in terms of additional linting and more complete documentation as the project envolves. In particular I want to go through all the items in the MA coding principles etc. as a separate issue.

Implements #1.