Adding evaulation step
Adds an evaluation class that holds the result of whether a layer for a given operation passes, fails, etc.

The evaluation each of these classes performs is based on the errors each layer has, which are mapped to an evaluation result.

Currently the control methods to setup these evaluations, and then to perform/execute them are separate but it may make more sense to combine them.