Add export step
Split into two steps, one to generate a dict with data structured for use in exporters, and a very minimal JSON exporter to demonstrate what this structure looks like as a file.

Note: This code is beyond the bits I was meant to do and so needs to be reviewed, especially given this will be the main interface exposed by this project (with the idea we may have multiple exporters). I added this largely so I could test everything was processed correctly, I’m not wedded to any of it.