Adding fixes for operations without any layers
To workaround new crash move folders where MapChef hasn’t yet run, and so can’t be evaluated.

As it turns out, this is only a problem for generating exports of dashboard data, which isn’t part of this app yet (it’s part of the undecided export step). However as it’s been a faff to get sort this repo out to create a clean commit, I’m not going to reverse it now.