Refactoring project layout
Add src directory containing a package for the application.

This is a more recommended way to develop Python packages to avoid differences between how this code behaves when used as a local module, vs when it’s installed as a package.

This change also refactors the `app_version` variable to use a conventional default name (`__version__`) and location (root of the package).