Adding minimal entrypoint
Now that we have an application package, we won’t use Poetry and so won’t have the ability to run the app via a script.

Instead we can run a module directly using `python -m` which will work as a way to run the app until we support alternative ways to do this (e.g. using a workflow engine).